Feature/#80 : 큐레이션 좋아요 기능 제거, 네비게이션 개선, KeywordChip 및 키워드 상세 화면 추가

+++ feature/login/src/main/java/com/example/login/viewmodel/LoginViewModel.kt
@@ -98,7 +98,7 @@ open class LoginViewModel @Inject constructor(
 
                 Log.d(TAG, "로그인 성공")
 
-                val userId = result.userId.toLong()
+                val userId = result.userId
 
                 // 토큰 + userId 저장
                 authPreference.saveTokens(
